Bug fix of "factorial" function: Can now calculate a factorial of up to  n < 21 (required for EGM96)

+++ src/util/InsMath.hpp
@@ -38,6 +38,6 @@ double pitchFromStaticAccelerationObs(const Eigen::Vector3d& accel_b);
 /// @brief Calculates the factorial of an unsigned integer
 /// @param[in] n Unsigned integer
 /// @return The factorial
-uint32_t factorial(uint32_t n);
+uint64_t factorial(uint64_t n);
 
 } // namespace NAV
